% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{all_genes}
\alias{all_genes}
\title{All genes description table}
\format{
A \code{tibble} object with 24488 rows and 47 columns.
\itemize{
\item Rows correspond to CT genes
\item Columns give CT genes characteristics
}
}
\source{
See \code{scripts/make_all_genes.R} in \code{CTdata} for details on how this
list was created.
}
\usage{
all_genes
}
\value{
A tibble of all 24 488 genes with their characteristics
}
\description{
All genes description, imported from \code{CTdata}
}
\details{
See \code{CTdata::all_genes} documentation for details
}
\examples{
all_genes

}
\keyword{datasets}
