% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCLE_correlation_matrix}
\alias{CCLE_correlation_matrix}
\title{CCLE correlation_matrix}
\format{
A \code{matrix} object with 306 rows and 24327 columns.
\itemize{
\item Rows correspond to CT genes
\item Rows correspond to all genes from CCLE database
}
}
\source{
(see inst/scripts/make_CCLE_correlation_matrix.R for details)
}
\usage{
CCLE_correlation_matrix
}
\description{
CCLE correlation matrix
}
\details{
Correlation coefficients between CT genes and all other genes
are given in the matrix. These correlation coefficients were calculated
from using log transformed expression values from \code{CCLE_data}.
}
\keyword{datasets}
