% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gene_names.R
\name{check_gene_names}
\alias{check_gene_names}
\title{Keeps genes present in a specified SE}
\usage{
check_gene_names(genes, database)
}
\arguments{
\item{genes}{Genes}

\item{database}{\code{SE} The summarizeExperiment must contain a column
external_gene_name in it rowData.}
}
\value{
genes present in the specified database
}
\description{
Checks if specified genes are present in the rowData's column
\code{external_gene_name} of a SummarizedExperiment, and returns the genes that
are in it.
}
\examples{
check_gene_names(genes = c("MAGEA1", "MAGEA3", "XXX"), database = GTEX_data)
}
