% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_genes.R
\name{correlated_genes}
\alias{correlated_genes}
\title{Gene correlations in CCLE cancer cell lines}
\usage{
correlated_genes(gene, corr_thr = 0.5, return = FALSE)
}
\arguments{
\item{gene}{CT gene selected}

\item{corr_thr}{Genes with an absolute correlation coefficient (Pearson)
higher than this threshold will be highlighted (default = 0.5)}

\item{return}{Boolean (FALSE by default). If set to TRUE, the function will
return the correlation coefficients with all genes instead of the plot.}
}
\value{
A plot where each dots represent the correlation coefficients (Pearson)
between genes and the specified CT gene (entered as input). Genes with a
correlation coefficient above threshold are colored in red if they are CT
genes or in blue, if not. If return = TRUE, all correlations coefficients are
returned instead.
}
\description{
A function that uses expression data from CCLE cell lines and
highlights genes correlated (or anti-correlated) with specified CT gene.
Genes with a correlation coefficient above threshold are colored in red if
they are CT genes or in blue, if not.
}
\examples{
correlated_genes(gene = "MAGEA3")
correlated_genes("TDRD1", 0.3)
}
