% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CT_mean_methylation_in_tissues}
\alias{CT_mean_methylation_in_tissues}
\title{Promoter mean methylation of Cancer Testis Genes in normal tissues}
\format{
A SummarizedExperiment object with 306 rows and 14 columns
\itemize{
\item Rows correspond to CT genes
\item Mean methylation levels in normal tissues are stored in columns
\item CpG densities and results of methylation analysis are stored
in rowData
}
}
\source{
WGBS methylation data was downloaded from Encode and from GEO
databases. Mean methylation levels are evaluated using methylation values
of CpG located in promoter region (defined as 1000 nt upstream TSS and
200 nt downstream TSS) (see inst/scripts/make_CT_mean_methylation_in_tissues.R
for details)
}
\usage{
CT_mean_methylation_in_tissues
}
\description{
Mean methylation values of CpG located within Cancer-Testis
(CT) promoters in a set of normal tissues
}
\details{
The RowData contains:
\itemize{
\item A column named \code{CpG_density}, gives the density of CpG within each promoter
(number of CpG / promoter length * 100).
\item A column \code{CpG_promoter} that classifies the promoters according to their
CpG densities: "CpG_low" (CpG_density < 2), "CpG_intermediate"
(CpG_density >= 2 & CpG_density < 4), and "CpG_high (CpG_density >= 4).
\item A column \code{methylation_in_tissues} summarising the methylation analysis
in normal tissues.
}
}
\keyword{datasets}
