% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAC_induction.R
\name{DAC_induction}
\alias{DAC_induction}
\title{Gene expression in cells treated or not by a demethylating agent}
\usage{
DAC_induction(database, genes = NULL)
}
\arguments{
\item{database}{Can be \code{DAC_treated_cells} or\code{ DAC_treated_cells_multimapping},
depending if returned expression values should take into account or not
multi-mapped reads.}

\item{genes}{Genes selected (all CT genes by default)}
}
\value{
A heatmap of selected genes in cells treated or not by a demethylating
agent. Gene normalised logcounts are invisibly returned.
}
\description{
Plots a heatmap of gene normalised counts (log-transformed)
in a selection of cells treated or not by 5-Aza-2′-Deoxycytidine (DAC),
a demethylating agent.
}
\details{
RNAseq data from cells treated or not with 5-aza downloaded from SRA.
Data was processed using a standard RNAseq pipeline.
\href{https://ccb.jhu.edu/software/hisat2/index.shtml}{hisat2} was used to align
reads to grch38 genome.
\href{https://rdrr.io/bioc/Rsubread/man/featureCounts.html}{featurecounts} was used
to assign reads to genes. Note that -M parameter was used or not to allow or not
counting multi-mapping reads.
}
\examples{
DAC_induction(database = DAC_treated_cells, genes = c("MAGEA1", "MAGEA3",
"MAGEA4", "MAGEA6", "CTAG1A"))
DAC_induction(database = DAC_treated_cells_multimapping,
genes = c("MAGEA1", "MAGEA3", "MAGEA4", "MAGEA6", "CTAG1A"))
}
