% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DAC_treated_cells_multimapping}
\alias{DAC_treated_cells_multimapping}
\title{RNseq expression values (multi-mapping was allowed) and differential
expression analysis of genes in cell lines treated or not with a
demethylating agent.}
\format{
A SummarizedExperiment object with 24359 rows and 32 columns
\itemize{
\item Rows correspond to genes (ensembl_gene_id)
\item Columns correspond to samples
\item Expression data are normalised counts log-transformed (log1p)
\item The colData contains the SRA references from which fastq files
were downloaded
\item Results of a differential expression evaluating the DAC treatment
effect are stored in the rowData.
}
}
\source{
RNAseq fastq files were downloaded from Encode database.
SRA reference of samples are stored in the colData.
}
\usage{
DAC_treated_cells_multimapping
}
\description{
Gene expression values in a set of cell lines treated
or not with 5-Aza-2′-Deoxycytidine (DAC), a demethylating agent.
Many CT genes belong to gene families from which members have
identical or nearly identical sequences. Some CT can only be detected
in RNAseq data in which multimapping reads are not discared.
}
\details{
Differential expression analysis was done using DESeq2_1.36.0,
using as design = ~ cell + treatment
(see inst/scripts/make_DAC_treated_cells_multimapping.R for details).
}
\keyword{datasets}
