% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTEX_data}
\alias{GTEX_data}
\title{Genes expression in GTEX}
\format{
A \code{SummarizedExperiment} object with 24359 rows and 30 columns
\itemize{
\item Rows correspond to genes (ensembl_gene_id)
\item Columns correspond to tissues
\item Expression data from the assay are TPM values
}
}
\source{
Downloaded from
https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_median_tpm.gct.gz.
Some categories of tissues were pooled (mean expression values are
given in pooled tissues). (see inst/scripts/make_GTEX_data.R for details)
}
\usage{
GTEX_data
}
\description{
Gene expression data in normal tissues from GTEX database.
}
\details{
The RowData contains
\itemize{
\item A column named \code{GTEX_category}, specifying the tissue specificity
category ("testis_specific", "testis-preferential", "lowly_expressed" or
"other") assigned to each gene using expression values in testis and in
somatic tissues, has been added to the rowData.
\item A column named \code{TPM_testis} giving the expression level in testis.
\item A column named \code{max_TPM_somatic} giving the maximum expression level
found in a somatic tissue.
\item A column named \code{q75_TPM_somatic} giving the quantile 75\% expression level
found in a somatic tissue.
}
}
\keyword{datasets}
