% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEX_expression.R
\name{GTEX_expression}
\alias{GTEX_expression}
\title{Gene expression in normal tissues (GTEx)}
\usage{
GTEX_expression(database, genes = NULL, units = "TPM")
}
\arguments{
\item{database}{GTEX_data}

\item{genes}{Genes selected (all CT genes by default)}

\item{units}{Expression values unit.
Can be "TPM" (default) or "log_TPM" (log(TPM + 1))}
}
\value{
A heatmap of selected genes in normal tissues.
Expression values are invisibly returned.
}
\description{
Plots an expression heatmap of genes in normal
tissues (GTEx database)
}
\examples{
GTEX_expression(database = GTEX_data)
GTEX_expression(database = GTEX_data, genes = c("MAGEA1", "MAGEA3"),
units = "log_TPM")
}
