% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_tissue_expression_multimapping.R
\name{normal_tissue_expression_multimapping}
\alias{normal_tissue_expression_multimapping}
\title{Expression values (TPM) of genes in normal tissues with or without multimapping}
\usage{
normal_tissue_expression_multimapping(
  database,
  genes = NULL,
  multimapping = NULL,
  units = "TPM"
)
}
\arguments{
\item{database}{normal_tissues_multimapping_data}

\item{genes}{genes selected (all CT genes by default)}

\item{multimapping}{Set to TRUE or FALSE to specify if returned expression
values must take into account or not multi-mapped reads}

\item{units}{Expression values units.
Can be "TPM" (default) or "log_TPM" (log(TPM + 1))}
}
\value{
A heatmap of selected gene expression values (TPM) in a set of
normal tissues calculated by counting or discarding multi-mapped reads.
Gene TPM values are invisibly returned.
}
\description{
Plots a heatmap of gene expression values in a set of normal
tissues. Expression values (in TPM) have been evaluated by either counting or
discarding multi-mapped reads. Indeed, many CT genes belong to gene families
from which members have identical or nearly identical sequences. Some CT can
only be detected in RNAseq data in which multimapping reads are not discared.
}
\details{
RNAseq data from a set of normal tissues were downloaded from Encode.
(see inst/scripts/make_CT_normal_tissues_multimapping.R for fastq references)
Fastq files were processed using a standard RNAseq pipeline including
}
\examples{
normal_tissue_expression_multimapping(database = normal_tissues_multimapping_data,
genes = c("GAGE13", "CT45A6", "NXF2", "SSX2", "CTAG1A", "MAGEA3", "MAGEA6"),
multimapping = FALSE)
normal_tissue_expression_multimapping(database = normal_tissues_multimapping_data,
genes = c("GAGE13", "CT45A6", "NXF2", "SSX2", "CTAG1A", "MAGEA3", "MAGEA6"),
multimapping = TRUE)
}
