% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_tissues_mean_methylation.R
\name{normal_tissues_mean_methylation}
\alias{normal_tissues_mean_methylation}
\title{Promoter methylation of Cancer-Testis genes in normal tissues}
\usage{
normal_tissues_mean_methylation(
  database,
  genes = NULL,
  include_genes_with_missing_values = FALSE
)
}
\arguments{
\item{database}{CT_mean_methylation_in_tissues}

\item{genes}{name of CT gene selected}

\item{include_genes_with_missing_values}{Set to TRUE or FALSE to specify if
genes with missing methylation values in some tissues should be included
(set to FALSE by default). Note that no gene clustering will be done when
some methylation values are missing.}
}
\value{
heatmap of mean promoter methylation of Cancer-Testis (CT) genes
in normal tissues. Methylation values are returned invisibly.
}
\description{
Plots a heatmap of mean promoter methylation levels of
Cancer-Testis (CT) genes in normal tissues. Methylation levels in
tissues correspond to the mean methylation of CpGs located in range of
1000 pb upstream and 200 pb downstream from gene TSS.
}
\examples{
normal_tissues_mean_methylation(CT_mean_methylation_in_tissues)
normal_tissues_mean_methylation(CT_mean_methylation_in_tissues,
c("MAGEA1", "MAGEA2", "MAGEA3", "MAGEA4"))
normal_tissues_mean_methylation(CT_mean_methylation_in_tissues,
c("MAGEA1", "MAGEA2", "MAGEA3", "MAGEA4"),
include_genes_with_missing_values = TRUE)
}
