% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_tissues_methylation.R
\name{normal_tissues_methylation}
\alias{normal_tissues_methylation}
\title{Promoter methylation of Cancer-Testis genes in normal tissues}
\usage{
normal_tissues_methylation(database, gene, nt_up = NULL, nt_down = NULL)
}
\arguments{
\item{database}{CT_methylation_in_tissues}

\item{gene}{name of selected CT gene}

\item{nt_up}{number of nucleotides upstream the TSS to analyse (1000 by default)}

\item{nt_down}{number of nucleotides downstream the TSS to analyse (200 by default)}
}
\value{
heatmap of Promoter methylation of a Cancer-Testis (CT) gene in normal tissues.
Methylation values are returned invisibly.
}
\description{
Plots a heatmap of Promoter methylation of a Cancer-Testis (CT)
gene in normal tissues
}
\examples{
normal_tissues_methylation(CT_methylation_in_tissues, gene = "TDRD1", 1000, 0)
}
