% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_TCGA_methylation_expression}
\alias{prepare_TCGA_methylation_expression}
\title{Prepare methylation and expression data of a gene in TCGA tumors}
\usage{
prepare_TCGA_methylation_expression(
  tumor = "all",
  gene = NULL,
  nt_up = NULL,
  nt_down = NULL,
  include_normal_tissues = FALSE
)
}
\arguments{
\item{tumor}{\code{character} defining the TCGA tumor type. Can be one
of "SKCM", "LUAD", "LUSC", "COAD", "ESCA", "BRCA", "HNSC", or
"all" (default).}

\item{gene}{\code{character} selected CT gene.}

\item{nt_up}{\code{numeric(1)} indicating the number of nucleotides
upstream the TSS to define the promoter region (1000 by
default)}

\item{nt_down}{\code{numeric(1)} indicating the number of nucleotides
downstream the TSS to define the promoter region (200 by
default)}

\item{include_normal_tissues}{\code{logical(1)}. If \code{TRUE},
the function will include normal peritumoral tissues in addition
to tumoral samples. Default is \code{FALSE}.}
}
\value{
a Dataframe giving for each TCGA sample, the methylation level of
a gene (mean methylation of probes located in its promoter) and the
expression level of the gene (TPM value). The number of probes used to
estimate the methylation level is also reported.
}
\description{
Creates a Dataframe giving for each TCGA sample, the methylation level of
a gene (mean methylation of probes located in its promoter) and the
expression level of the gene (TPM value).
}
\examples{
CTexploreR:::prepare_TCGA_methylation_expression("LUAD", gene = "TDRD1")
}
