% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setFontSize}
\alias{setFontSize}
\title{Determine font size}
\usage{
setFontSize(matrix)
}
\arguments{
\item{matrix}{\code{matrix} containing the data to visualise}
}
\value{
A logical number that is the fontsize to use
}
\description{
Gives the fontsize to use for the heatmap based on the matrix's dimension.
}
\examples{
CTexploreR:::setFontSize(matrix(1:3, 9,8))
}
