% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_database}
\alias{subset_database}
\title{Subset databases}
\usage{
subset_database(variable = NULL, data, include_CTP = FALSE)
}
\arguments{
\item{variable}{\code{character()} containing the names genes to keep in the
data. The default value, \code{NULL}, takes CT (specific) genes}

\item{data}{\verb{Summarized Experiment} or \code{SingleCellExperiment} object
with valid variable names.}

\item{include_CTP}{\code{logical(1)} If \code{TRUE}, CTP genes are included.
(\code{FALSE} by default).}
}
\value{
A \verb{Summarized Experiment} or \code{SingleCellExperiment} object with
only the variables data
}
\description{
Check the presence of the genes in the database then subsets the database
to only keep these genes' data.
}
\examples{

  CTexploreR:::subset_database(variable = "MAGEA1", data = CTdata::GTEX_data())
}
