% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_methylation_expression_correlation.R
\name{TCGA_methylation_expression_correlation}
\alias{TCGA_methylation_expression_correlation}
\title{Methylation-Expression correlation of Cancer-Testis genes in TCGA samples}
\usage{
TCGA_methylation_expression_correlation(
  expression_database,
  methylation_database,
  tumor,
  gene = NULL,
  corr_coeff = FALSE
)
}
\arguments{
\item{expression_database}{TCGA_TPM}

\item{methylation_database}{TCGA_CT_methylation}

\item{tumor}{TCGA tumor code. c("SKCM", "LUAD", "LUSC", "COAD", "ESCA", "BRCA", "HNSC", "all")}

\item{gene}{CT gene name(s)}

\item{corr_coeff}{Boolean (FALSE by default). If set to TRUE, the function
will invisibly return the correlation coefficient (Pearson), between
methylation and expression values for the gene in selected samples.}
}
\value{
A correlation plot between gene expression and methylation
values of probe(s) located in its promoter region (defined as 1000
nucleotides upstream TSS and 200 nucleotides downstream TSS).
By default, methylation and expression values for the gene in
selected tumors are invisibly returned in a tibble. If \code{corr_coeff} is
set to TRUE, the correlation coefficient is be invisibly returned
instead of methylation and expression values.
}
\description{
Plots the correlation between methylation and
expression values of a Cancer-Testis (CT) gene in TCGA samples.
}
\details{
The coefficient of correlation is set to \code{NA} if no probes are
found in promoter regions or if less than 1\% of tumors are positive
(TPM >= 1) for the gene.
}
\examples{
TCGA_methylation_expression_correlation(TCGA_TPM, TCGA_CT_methylation,
tumor = "LUAD", gene = "TDRD1")
TCGA_methylation_expression_correlation(TCGA_TPM, TCGA_CT_methylation,
tumor = "all", gene = "TDRD1")
TCGA_methylation_expression_correlation(TCGA_TPM, TCGA_CT_methylation,
tumor = "all", gene = "TDRD1",
corr_coeff = TRUE)
}
