% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_TPM}
\alias{TCGA_TPM}
\title{Gene expression in TCGA samples}
\format{
A SummarizedExperiment object with 24350 rows and 4087 columns
\itemize{
\item Rows correspond to CT genes (ensembl reference)
\item Columns correspond to samples
\item Expression data from the assay are TPM values
\item Clinical information are stored in colData
\item Genes information are stored in RowData
}
}
\source{
SKCM, LUAD, LUSC, COAD, ESCA, BRCA and HNSC expression data were downloaded
with TCGAbiolinks. Global hypomethylation levels from paper \verb{DNA methylation loss promotes immune evasion of tumours with high mutation and copy number load}
from Jang et al., Nature Commun 2019 were added to colData.
(see inst/scripts/make_TCGA_TPM.R for details)
}
\usage{
TCGA_TPM
}
\description{
Gene expression data in TCGA samples (tumor and peritumoral samples)
}
\keyword{datasets}
