% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_names}
\alias{check_names}
\title{Check spelling of entered variables}
\usage{
check_names(variable, valid_vector)
}
\arguments{
\item{variable}{\code{character()} containing the names of variables to
check.}

\item{valid_vector}{\code{character()} with valid variable names.}
}
\value{
A character with valid variables.
}
\description{
Checks the spelling of a vector of entered variable(s) comparing it
to a vector of valid names, and removes the ones that are absent
from the vector of valid names.
}
\examples{
CTexploreR:::check_names(variable = c("Ovarian", "leukemia", "wrong_name"),
                         valid_vector = c("ovarian", "leukemia"))
}
