% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_genes.R
\name{correlated_genes}
\alias{correlated_genes}
\title{Gene correlations in CCLE cancer cell lines}
\usage{
correlated_genes(corr_matrix, gene, corr_thr = 0.5)
}
\arguments{
\item{corr_matrix}{CCLE_correlation_matrix}

\item{gene}{CT gene selected}

\item{corr_thr}{Genes with an absolute correlation coefficient higher
than this threshold will be highlighted (default = 0.5)}
}
\value{
A plot where each dots represent the correlation coefficients between
genes and the specified CT gene (entered as input). All correlations coefficients
are invisibly returned.
}
\description{
A function that uses expression data from CCLE cell lines and
highlights genes correlated (or anti-correlated) with specified CT gene.
}
\examples{
correlated_genes(CCLE_correlation_matrix, gene = "MAGEA3")
correlated_genes(CCLE_correlation_matrix, "TDRD1", 0.3)
}
