% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CT_genes}
\alias{CT_genes}
\title{CT genes description table}
\format{
A \code{tibble} object with 306 rows and 33 columns.
\itemize{
\item Rows correspond to CT genes
\item Columns give CT genes characteristics
}
}
\source{
(see inst/scripts/make_CT_genes.R for details)
}
\usage{
CT_genes
}
\description{
Cancer-Testis (CT) genes description
}
\details{
CT_genes characteristics column:
\itemize{
\item Column \code{family} gives the gene family name.
\item Column \code{X_linked} classifies genes in "chrX" or "not_chrX".
\item Column \code{GTEX_category} gives the category ("testis_specific",
"testis_preferential" or "lowly_expressed") assigned to each gene
using GTEx database.
\item Column \code{TPM_testis} gives the gene expression level in testis
(using GTEx database).
\item Column \code{max_TPM_somatic} gives the maximum expression level
found in a somatic tissue (using GTEx database).
\item Column \code{q75_TPM_somatic} gives the 75\% quantile expression level
found in a somatic tissue (using GTEx database).
\item Column \code{multimapping_analysis} informs if the gene was found to
be testis-specific when multi-mapped reads were counted for gene
expression in normal tissues ("not_analysed" or "testis_specific").
\item Column \code{testis_specificity} gives the testis-specificity of genes
assigned to each gene using \code{GTEX_category} and \code{multimapping_analysis}
("testis_specific" or "testis_preferential").
\item Column \code{percent_of_positive_CCLE_cell_lines} gives the percentage of
CCLE cancer cell lines in which genes are expressed (genes were
considered as expressed if TPM >= 10).
\item Column \code{percent_of_negative_CCLE_cell_lines} gives the percentage of
CCLE cancer cell lines in which genes are repressed (TPM <= 0.1).
\item Column \code{max_TPM_in_CCLE} gives the highest expression level of genes
in CCLE cell lines.
\item Column \code{CCLE_category} gives the category assigned to each gene
using CCLE data. "Activated" category corresponds to genes expressed
in at least one cell line (TPM >= 10) and repressed in at least 20\% of
cell lines.
\item Column \code{percent_pos_tum} gives the percentage of TCGA cancer samples in
which genes are expressed (genes were considered as expressed if TPM >= 10).
\item Column \code{percent_neg_tum} gives the percentage of TCGA cancer samples in
which genes are repressed (TPM <= 0.1).
\item Column \code{max_TPM_in_TCGA} gives the highest expression level of genes
in TCGA cancer sample.
\item Column \code{TCGA_category} gives the category assigned to each gene
using TCGA data. "activated" category corresponds to genes expressed
in at least one tumor (TPM >= 10) and repressed in at least 20\% of samples.
\item Column \code{DAC} summarises the results ("induced" or "not_induced") of a
differential expression evaluating gene induction upon DAC treatment
in a series of cell lines.
\item Column \code{methylation_in_tissues} summarises the analysis of gene
promoter methylation in somatic and germline normal tissues
("methylated_in_somatic_unmethylated_in_germline",
"methylated_in_somatic_and_germline" or "unmethylated_in_somatic").
\item Column \code{regulation} summarises the regulation category ("methylation"
or "not_methylation") that was assigned to genes based on DAC induction
and on promoter methylation levels in normal tissues (when available).
\item Column \code{met_exp_corr_TCGA} gives the correlation coefficient between
promoter methylation and expression in TCGA samples.
\item Column named \code{CpG_density}, gives the density of CpG within each promoter
(number of CpG / promoter length * 100).
\item Column \code{CpG_promoter} classifies the promoters according to their
CpG densities: "CpG_low" (CpG_density < 2), "CpG_intermediate"
(CpG_density >= 2 & CpG_density < 4), and "CpG_high" (CpG_density >= 4).
\item Columns \code{external_transcript_name}, \code{ensembl_transcript_id},
\code{chromosome_name}, \code{strand}, \code{transcription_start_site},
\code{transcript_length} and \code{transcript_biotype} give the references and
informations about the most biologically relevant transcript associated to
each gene.
\item Columns \code{oncogene} and \code{tumor_suppressor} informs if oncogenic and
tumor-suppressor functions have been associated to genes
(source: \href{http://bionlp.bcgsc.ca/cancermine/}{Cancermine}).
}
}
\keyword{datasets}
