% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{CT_genes}
\alias{CT_genes}
\title{CT genes description table}
\format{
A \code{tibble} object with 308 rows and 34 columns.
\itemize{
\item Rows correspond to CT genes
\item Columns give CT genes characteristics
}
}
\source{
See \code{scripts/make_CT_genes.R} in \code{CTdata} for details on how this list of
curated CT genes was created.
}
\description{
Cancer-Testis (CT) genes description, imported from \code{CTdata}
}
\details{
See \code{CTdata::CT_genes} documentation for details
}
