% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_tissues_methylation.R
\name{normal_tissues_methylation}
\alias{normal_tissues_methylation}
\title{Methylation of CpGs located in Cancer-Testis promoters in normal tissues}
\usage{
normal_tissues_methylation(gene, nt_up = NULL, nt_down = NULL, return = FALSE)
}
\arguments{
\item{gene}{Name of selected CT gene}

\item{nt_up}{Number of nucleotides upstream the TSS to analyse
(1000 by default)}

\item{nt_down}{Number of nucleotides downstream the TSS to analyse
(200 by default)}

\item{return}{Boolean (FALSE by default). If set to TRUE, the function will
return the methylation values of all cytosines in the promoter instead of
the heatmap.}
}
\value{
Heatmap of the methylation of CpGs located in a Cancer-Testis (CT)
promoter, in normal tissues. If return = TRUE, methylation values are
returned instead.
}
\description{
Plots a heatmap of the methylation of CpGs located in a
Cancer-Testis (CT) promoter, in normal tissues. X-axis corresponds to the
CpGs position (related to TSS).
}
\examples{
normal_tissues_methylation(gene = "TDRD1", 1000, 0)
}
