% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_database}
\alias{subset_database}
\title{Subset databases}
\usage{
subset_database(variable, data)
}
\arguments{
\item{variable}{\code{character()} containing the names genes to keep in the
data}

\item{data}{\verb{Summarized Experiment} or \code{SingleCellExperiment} object
with valid variable names.}
}
\value{
A \verb{Summarized Experiment} or \code{SingleCellExperiment} object with
only the variables data
}
\description{
Check the presence of the genes in the database then subsets the database
to only keep these genes' data.
}
\examples{
CTexploreR:::subset_database(variable = "MAGEA1", data = CTdata::GTEX_data())
}
