% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_expression.R
\name{TCGA_expression}
\alias{TCGA_expression}
\title{Gene expression in TCGA tumors}
\usage{
TCGA_expression(database, tumor = "all", genes = NULL, units = "TPM")
}
\arguments{
\item{database}{TCGA_TPM}

\item{tumor}{TCGA tumor code. can be one of "SKCM", "LUAD", "LUSC", "COAD", "ESCA",
"BRCA", "HNSC", or "all" (default).}

\item{genes}{Genes selected (All CT genes by default)}

\item{units}{Expression values unit.
Can be "TPM" (default) or "log_TPM" (log(TPM + 1))}
}
\value{
A heatmap of selected CT genes in TCGA samples.
A SummarizedExperiment with TPM expression data is invisibly returned.
}
\description{
Plots a heatmap of genes expression in TCGA samples
(peritumoral and tumor samples when a specific tumor type is specified, or
tumor samples only when tumor option is set to "all")
}
\examples{
TCGA_expression(TCGA_TPM, tumor = "LUAD", genes = c("MAGEA1", "MAGEA3", "MAGEA4"), units = "log_TPM")
}
