% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_methylation_expression_correlation.R
\name{TCGA_methylation_expression_correlation}
\alias{TCGA_methylation_expression_correlation}
\title{Methylation-Expression correlation of Cancer-Testis genes in TCGA samples}
\usage{
TCGA_methylation_expression_correlation(
  tumor,
  gene = NULL,
  nt_up = 1000,
  nt_down = 200,
  return = FALSE,
  corr_coeff = FALSE
)
}
\arguments{
\item{tumor}{TCGA tumor code. c("SKCM", "LUAD", "LUSC", "COAD", "ESCA",
"BRCA", "HNSC", "all")}

\item{gene}{CT gene name}

\item{nt_up}{Number of nucleotides upstream the TSS to define the promoter
region (1000 by default)}

\item{nt_down}{Number of nucleotides downstream the TSS to define the
promoter region (200 by default)}

\item{return}{Boolean (FALSE by default). If set to TRUE, the function will
return the methylation and expression values in all samples instead of the
heatmap.}

\item{corr_coeff}{Boolean (FALSE by default). If set to TRUE, the function
will invisibly return the correlation coefficient (Pearson), between
methylation and expression values for the gene in selected samples.}
}
\value{
A correlation plot between gene expression and methylation
values of probe(s) located in its promoter region (defined as 1000
nucleotides upstream TSS and 200 nucleotides downstream TSS).
If return = TRUE, methylation and expression values for the gene in
selected tumors are returned in a tibble instead. If \code{corr_coeff} is
set to TRUE, the correlation coefficient is being returned instead.
}
\description{
Plots the correlation between methylation and
expression values of a Cancer-Testis (CT) gene in TCGA samples.
}
\details{
The coefficient of correlation is set to \code{NA} if no probes are
found in promoter regions or if less than 1\% of tumors are positive
(TPM >= 1) for the gene.
}
\examples{
TCGA_methylation_expression_correlation("LUAD", gene = "TDRD1",
return = FALSE)
TCGA_methylation_expression_correlation(c("LUAD", "LUSC"),
gene = "MAGEA1", return = FALSE)
TCGA_methylation_expression_correlation("LUAD",
gene = "TDRD1", return = TRUE)
}
